AuthToken cleanup

+++ trunk/NoMatterWebApi/NoMatterWebApi/Controllers/V1/CategoryController.cs
@@ -1,35 +1,29 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net;
-using System.Net.Http;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using System.Web.Http;
 using System.Web.Http.Description;
 using NoMatterDatabaseModel;
 using NoMatterWebApi.ActionResults;
+using NoMatterWebApi.CustomAuth;
 using NoMatterWebApi.DAL;
 using NoMatterWebApi.Extensions;
 using NoMatterWebApi.Helpers;
 using NoMatterWebApi.Logging;
-using NoMatterWebApi.Models;
-using NoMatterWebApiModels.Models;
 using Category = NoMatterWebApiModels.Models.Category;
 using Product = NoMatterWebApiModels.Models.Product;
-using Section = NoMatterWebApiModels.Models.Section;
+
 
 namespace NoMatterWebApi.Controllers.V1
 {
 	//[RoutePrefix("api/v1/categories")]
 	public class CategoryController : ApiController
 	{
-		private IUserRepository _userRepository;
 		private IProductRepository _productRepository;
-		private ISectionRepository _sectionRepository;
 		private ICategoryRepository _categoryRepository;
-		private IClientRepository _clientRepository;
 		private IGeneralHelper _generalHelper;
-		private IWebApiGlobalSettings _webApiGlobalSettings;
 		
 
 		public CategoryController()
@@ -38,35 +32,26 @@ namespace NoMatterWebApi.Controllers.V1
 
 			_productRepository = new ProductRepository(databaseEntity);
 			_categoryRepository = new CategoryRepository(databaseEntity);
-			_sectionRepository = new SectionRepository(databaseEntity);
-			_userRepository = new UserRepository(databaseEntity);
-			_clientRepository = new ClientRepository(databaseEntity);
 			_generalHelper = new WebApiGeneralHelper();
-			_webApiGlobalSettings = new WebApiGlobalSettings();
 			
 		}
 
-		public CategoryController(IProductRepository productRepository, ICategoryRepository categoryRepository, ISectionRepository sectionRepository, IClientRepository clientRepository, IGeneralHelper generalHelper)
+		public CategoryController(IProductRepository productRepository, ICategoryRepository categoryRepository, IGeneralHelper generalHelper)
 		{
 			_productRepository = productRepository;
-			_categoryRepository = categoryRepository;
-			_sectionRepository = sectionRepository;
-			_clientRepository = clientRepository;
+			_categoryRepository = categoryRepository;;
 			_generalHelper = generalHelper;
 		}
 
-		
 	
-		// GET api/v1/category/{categoryid}
-		[System.Web.Http.HttpGet]
-		[System.Web.Http.Route("api/v1/clients/{clientId}/categories/{categoryId}")]
+		[HttpGet]
+		[Route("api/v1/clients/{clientId}/categories/{categoryId}")]
 		[ResponseType(typeof(Category))]
 		public async Task<IHttpActionResult> GetCategoryAsync(string clientId, string categoryid)
 		{
 			try
 			{
-				//TODO: make sure the user can access this category
-				//Need to get bearer token, and lookup the user so we know which client the user is from
+				//TODO: make sure the category is part of this client
 
 				var categoryDb = await _categoryRepository.GetCategoryAsync(new Guid(clientId), new Guid(categoryid));
 
@@ -84,28 +69,33 @@ namespace NoMatterWebApi.Controllers.V1
 			}
 		}
 
-		// POST api/v1/sections
-		[System.Web.Http.HttpPut]
-		[System.Web.Http.Authorize]
-		[System.Web.Http.Route("api/v1/clients/{clientId}/categories/{categoryId}")]
+		[HttpPut]
+		[Authorize]
+		[Route("api/v1/clients/{clientId}/categories/{categoryId}")]
 		public async Task<IHttpActionResult> UpdateCategoryAsync(string clientId, string categoryId, Category category)
 		{
-			//TODO: make sure the user can post to this category
-
 			try
 			{
-				var userToken = User.Identity.Name;
+				//Make sure the user has access to this client
+				var claimsPrincipal = (ClaimsPrincipal)User;
+				var authUserClientId = claimsPrincipal.FindFirst(x => x.Type == CustomAuthentication.ClientId).Value;
 
-				var userDb = await _userRepository.GetClientUserByTokenAsync(userToken);
-				if (userDb == null) return new CustomBadRequest(Request, ApiResultCode.UserNotFound);
+				if (!string.IsNullOrEmpty(authUserClientId) && clientId != authUserClientId)
+					return new CustomBadRequest(Request, ApiResultCode.UserDoesNotBelongToClient);
 
-				var clientDb = await _clientRepository.GetClientAsync(new Guid(clientId));
-				if (clientDb == null) return new CustomBadRequest(Request, ApiResultCode.ClientNotFound);
+
+				//var userToken = User.Identity.Name;
+
+				//var userDb = await _userRepository.GetClientUserByTokenAsync(userToken);
+				//if (userDb == null) return new CustomBadRequest(Request, ApiResultCode.UserNotFound);
+
+				//var clientDb = await _clientRepository.GetClientAsync(new Guid(clientId));
+				//if (clientDb == null) return new CustomBadRequest(Request, ApiResultCode.ClientNotFound);
 
 				var categoryDb = await _categoryRepository.GetCategoryAsync(new Guid(clientId), new Guid(categoryId));
 				if (categoryDb == null) return new CustomBadRequest(Request, ApiResultCode.CategoryNotFound);
 
-				if (userDb.Client != null && userDb.ClientId != clientDb.ClientId) return new CustomBadRequest(Request, ApiResultCode.UserDoesNotBelongToClient);
+				//if (userDb.Client != null && userDb.ClientId != clientDb.ClientId) return new CustomBadRequest(Request, ApiResultCode.UserDoesNotBelongToClient);
 
 				//Update the section
 				await _categoryRepository.UpdateCategoryAsync(categoryDb, category);
@@ -120,15 +110,21 @@ namespace NoMatterWebApi.Controllers.V1
 			}
 		}
 
-		// DELETE api/v1/categories/{categoryid}
-		[System.Web.Http.HttpDelete]
-		[System.Web.Http.Route("api/v1/clients/{clientId}/categories/{categoryId}")]
-		public async Task<IHttpActionResult> DeleteCategoryAsync(string categoryId)
-		{
-			//TODO: make sure the user can delete this category
 
+		[Authorize]
+		[HttpDelete]
+		[Route("api/v1/clients/{clientId}/categories/{categoryId}")]
+		public async Task<IHttpActionResult> DeleteCategoryAsync(string clientId, string categoryId)
+		{
 			try
 			{
+				//Make sure the user has access to this client
+				var claimsPrincipal = (ClaimsPrincipal)User;
+				var authUserClientId = claimsPrincipal.FindFirst(x => x.Type == CustomAuthentication.ClientId).Value;
+
+				if (!string.IsNullOrEmpty(authUserClientId) && clientId != authUserClientId)
+					return new CustomBadRequest(Request, ApiResultCode.UserDoesNotBelongToClient);
+
 				await _categoryRepository.DeleteCategoryAsync(new Guid(categoryId));
 
 				return Ok();
@@ -140,13 +136,9 @@ namespace NoMatterWebApi.Controllers.V1
 			}
 		}
 
-		
-
-
-
-		// GET api/v1/category/{categoryid}/products
-		[System.Web.Http.HttpGet]
-		[System.Web.Http.Route("api/v1/clients/{clientId}/categories/{categoryid}/products")]
+	
+		[HttpGet]
+		[Route("api/v1/clients/{clientId}/categories/{categoryid}/products")]
 		[ResponseType(typeof(List<Product>))]
 		public async Task<IHttpActionResult> GetCategoryProducts(string clientId, string categoryid)
 		{
@@ -163,7 +155,7 @@ namespace NoMatterWebApi.Controllers.V1
 
 				var products = productsDb.Select(x => x.ToDomainProduct()).ToList();
 
-				//If its he Sale category.. filter out sale items
+				//If its the Sale category.. filter out sale items
 				if (categoryDb.ActionName != null && categoryDb.ActionName.ToLower() == "sale")
 				{
 					products = products.Where(x => x.DiscountDetails.Discounted).ToList();
@@ -178,21 +170,25 @@ namespace NoMatterWebApi.Controllers.V1
 			}
 		}
 
-		// POST api/v1/category/{categoryid}/products
-		[System.Web.Http.HttpPost]
-		[System.Web.Http.Route("api/v1/clients/{clientId}/categories/{categoryid}/products")]
+		[Authorize]
+		[HttpPost]
+		[Route("api/v1/clients/{clientId}/categories/{categoryId}/products")]
 		[ResponseType(typeof(Product))]
-		public async Task<IHttpActionResult> AddCategoryProductAsync(string clientId, string categoryid, Product model)
+		public async Task<IHttpActionResult> AddCategoryProductAsync(string clientId, string categoryId, Product model)
 		{
-			//TODO: make sure the user can post to this category
-
 			try
 			{
-				var userToken = User.Identity.Name;
+
+				//Make sure the user has access to this client
+				var claimsPrincipal = (ClaimsPrincipal)User;
+				var authUserClientId = claimsPrincipal.FindFirst(x => x.Type == CustomAuthentication.ClientId).Value;
+
+				if (!string.IsNullOrEmpty(authUserClientId) && clientId != authUserClientId)
+					return new CustomBadRequest(Request, ApiResultCode.UserDoesNotBelongToClient);
 
 				var productUuid = Guid.NewGuid();
 
-				var categoryDb = await _categoryRepository.GetCategoryAsync(new Guid(clientId), new Guid(categoryid));
+				var categoryDb = await _categoryRepository.GetCategoryAsync(new Guid(clientId), new Guid(categoryId));
 
 				if (categoryDb == null) return new CustomBadRequest(Request, ApiResultCode.CategoryNotFound);
